requirements.txt
astroid==1.5.3
Babel==2.4.0
bcrypt==3.1.3
blinker==1.4
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
colorama==0.3.9
decorator==4.1.1
defusedxml==0.5.0
http://github.com/pythonforfacebook/facebook-sdk/tarball/master
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-BabelEx==0.9.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OAuth==0.12
flask-paginate==0.5.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Social==1.6.2
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
future==0.16.0
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauth2==1.9.0.post1
oauth2client==4.1.2
oauthlib==2.0.2
passlib==1.7.1
pbr==3.1.1
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyJWT==1.5.2
pylint==1.7.2
python-dotenv==0.6.5
python-twitter==3.3
python3-openid==3.1.0
pytz==2017.2
requests==2.18.3
requests-oauthlib==0.8.0
rsa==3.4.2
selenium==3.4.3
six==1.10.0
social-auth-core==1.4.0
social-auth-storage-sqlalchemy==1.1.0
speaklater==1.3
SQLAlchemy==1.1.11
sqlalchemy-migrate==0.11.0
sqlparse==0.2.3
Tempita==0.5.2
uritemplate==3.0.0
urllib3==1.18.1
waitress==1.0.2
Werkzeug==0.12.2
wrapt==1.10.10
WTForms==2.1